Rename ApativeHuffmanTree sourcecode filenames
