+++ OP2Utility.vcxproj.filters
@@ -39,9 +39,6 @@
     <ClInclude Include="src\Archives\CompressionType.h">
       <Filter>Archives</Filter>
     </ClInclude>
-    <ClInclude Include="src\Archives\AdaptHuffTree.h">
-      <Filter>Archives</Filter>
-    </ClInclude>
     <ClInclude Include="src\Archives\ArchiveFile.h">
       <Filter>Archives</Filter>
     </ClInclude>
@@ -129,6 +126,9 @@
     <ClInclude Include="src\Archives\BitStreamReader.h">
       <Filter>Archives</Filter>
     </ClInclude>
+    <ClInclude Include="src\Archives\AdaptiveHuffmanTree.h">
+      <Filter>Archives</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\XFile.cpp">
@@ -139,9 +139,6 @@
     </ClCompile>
     <ClCompile Include="src\ResourceManager.cpp">
       <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="src\Archives\AdaptHuffTree.cpp">
-      <Filter>Archives</Filter>
     </ClCompile>
     <ClCompile Include="src\Archives\ClmFile.cpp">
       <Filter>Archives</Filter>
@@ -183,6 +180,9 @@
       <Filter>Streams</Filter>
     </ClCompile>
     <ClCompile Include="src\Archives\BitStreamReader.cpp">
+      <Filter>Archives</Filter>
+    </ClCompile>
+    <ClCompile Include="src\Archives\AdaptiveHuffmanTree.cpp">
       <Filter>Archives</Filter>
     </ClCompile>
   </ItemGroup>
